+++ tools/selection/text/kpToolText_KeyboardEvents_HandleArrowKeys.cpp
@@ -63,7 +63,7 @@ void kpToolText::handleUpKeyPress (QKeyEvent *e,
     if (!textLines.isEmpty () && cursorRow > 0)
     {
         cursorRow--;
-        cursorCol = qMin (cursorCol, (int) textLines [cursorRow].length ());
+        cursorCol = qMin (cursorCol, textLines [cursorRow].length ());
         viewManager ()->setTextCursorPosition (cursorRow, cursorCol);
     }
 
@@ -81,10 +81,10 @@ void kpToolText::handleDownKeyPress (QKeyEvent *e,
     if (hasBegunShape ())
         endShape (currentPoint (), normalizedRect ());
 
-    if (!textLines.isEmpty () && cursorRow < (int) textLines.size () - 1)
+    if (!textLines.isEmpty () && cursorRow < textLines.size () - 1)
     {
         cursorRow++;
-        cursorCol = qMin (cursorCol, (int) textLines [cursorRow].length ());
+        cursorCol = qMin (cursorCol, textLines [cursorRow].length ());
         viewManager ()->setTextCursorPosition (cursorRow, cursorCol);
     }
